TheUser rework
Przepraszam ;^)
